requirements.txt
openai-whisper
langchain
langchain_google_genai
pyaudio
python-dotenv
librosa
pywhatkit